+++ management/model_application_saml_all_of_sp_verification_certificates.go
@@ -17,15 +17,16 @@ import (
 // ApplicationSAMLAllOfSpVerificationCertificates struct for ApplicationSAMLAllOfSpVerificationCertificates
 type ApplicationSAMLAllOfSpVerificationCertificates struct {
 	// An array that specifies the certificate IDs used to verify the service provider signature.
-	Id *string `json:"id,omitempty"`
+	Id string `json:"id"`
 }
 
 // NewApplicationSAMLAllOfSpVerificationCertificates instantiates a new ApplicationSAMLAllOfSpVerificationCertificates object
 // This constructor will assign default values to properties that have it defined,
 // and makes sure properties required by API are set, but the set of arguments
 // will change when the set of required properties is changed
-func NewApplicationSAMLAllOfSpVerificationCertificates() *ApplicationSAMLAllOfSpVerificationCertificates {
+func NewApplicationSAMLAllOfSpVerificationCertificates(id string) *ApplicationSAMLAllOfSpVerificationCertificates {
 	this := ApplicationSAMLAllOfSpVerificationCertificates{}
+	this.Id = id
 	return &this
 }
 
@@ -37,41 +38,33 @@ func NewApplicationSAMLAllOfSpVerificationCertificatesWithDefaults() *Applicatio
 	return &this
 }
 
-// GetId returns the Id field value if set, zero value otherwise.
+// GetId returns the Id field value
 func (o *ApplicationSAMLAllOfSpVerificationCertificates) GetId() string {
-	if o == nil || o.Id == nil {
+	if o == nil {
 		var ret string
 		return ret
 	}
-	return *o.Id
+
+	return o.Id
 }
 
-// GetIdOk returns a tuple with the Id field value if set, nil otherwise
+// GetIdOk returns a tuple with the Id field value
 // and a boolean to check if the value has been set.
 func (o *ApplicationSAMLAllOfSpVerificationCertificates) GetIdOk() (*string, bool) {
-	if o == nil || o.Id == nil {
+	if o == nil {
 		return nil, false
 	}
-	return o.Id, true
+	return &o.Id, true
 }
 
-// HasId returns a boolean if a field has been set.
-func (o *ApplicationSAMLAllOfSpVerificationCertificates) HasId() bool {
-	if o != nil && o.Id != nil {
-		return true
-	}
-
-	return false
-}
-
-// SetId gets a reference to the given string and assigns it to the Id field.
+// SetId sets field value
 func (o *ApplicationSAMLAllOfSpVerificationCertificates) SetId(v string) {
-	o.Id = &v
+	o.Id = v
 }
 
 func (o ApplicationSAMLAllOfSpVerificationCertificates) MarshalJSON() ([]byte, error) {
 	toSerialize := map[string]interface{}{}
-	if o.Id != nil {
+	if true {
 		toSerialize["id"] = o.Id
 	}
 	return json.Marshal(toSerialize)